Improve task reordering
Improve logic for reordering tasks on drag & drop.

One reordering always runs 2 database queries.

Creating a new task places it at the top of the status.

Changing the status moves the task at the top of the new status.

This does not change the logic around archiving and unarchiving since it will be changed later.

Bug: 189993006

Change-Id: I617ae57adeea3520b852ac7cead38bbf33ea8b5e