Convert Add/Edit Task fragment to a dialog
This UI is shown as a dialog on large screens, but on normal (phone)
screens it is shown as a regular fragment. Since we cannot change the
navigation destination type between configurations, it is implemented
as a DialogFragment whose window uses either the full screen on smaller
devices or part of the screen on larger devices.

Bug: 195357852
Change-Id: Ic79f1d008ce70349f78d7bbf9a358b533a874806